server/requirements.txt
amqp==5.1.1
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
billiard==4.1.0
black==23.3.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
Django==4.2.3
django-celery-results==2.5.1
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-filter==22.1
django-js-asset==2.1.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
html5lib==1.1
idna==3.4
kombu==5.3.1
l18n==2021.3
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pillow-heif==0.12.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telepath==0.3.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wagtail==5.0.2
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
Willow==1.5.2
